backend/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0

# Database
motor>=3.6.0
pymongo>=4.10.0

# Redis
redis>=5.2.0

# Kafka
aiokafka>=0.12.0

# Validation & Settings
pydantic>=2.10.0
pydantic-settings>=2.7.0
email-validator>=2.2.0

# AI / Computer Vision
mediapipe>=0.10.14
numpy>=1.26.0
Pillow>=10.0.0

# Logging
structlog>=24.4.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.25.0
httpx>=0.28.0

# Linting
ruff>=0.8.0
